requirements.txt
# 项目依赖文件 requirements.txt
# 功能： 列出项目所需的Python包及其版本
# 作者： Xu Guangyuan
# 日期： 2025/08/09

torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.2
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
pyyaml==6.0
tqdm==4.65.0
Pillow==9.5.0
opencv-python==4.7.0.72